Add build job in ci.yml

+++ Sort/BubbleSort/bubble_sort.h
@@ -1,6 +1,8 @@
 #ifndef BUBBLE_SORT_H
 #define BUBBLE_SORT_H
 
-void bubbleSort(int arr[], const int size);
+#include <cstddef>
+
+void bubbleSort(int arr[], const size_t size);
 
 #endif // BUBBLE_SORT_H
